+++ .gitignore
@@ -1,3 +1,9 @@
+# SveltKit specific files
+.DS_Store
+/build
+/.svelte-kit
+/package
+
 # Logs
 logs
 *.log
