+++ Changes.md
@@ -1,0 +1,5 @@
+## Changes
+
+- Changed utils to internal (kept dateUtils)
+- Added a homepage and a seperate docs page for more flexibility
+- Added types (not complete yet)
