--- docs/index.js
@@ -1,6 +1,0 @@
-import Docs from './docs.svelte';
-
-new Docs({
-  target: document.body,
-  props: {}
-});
