--- index.js
@@ -1,5 +1,0 @@
-import SveltyPicker, { config } from './src/SveltyPicker.svelte';
-
-export default SveltyPicker;
-
-export { config }
