+++ src/app.css
@@ -1,0 +1,37 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+html {
+	tab-size: 2;
+}
+
+/** 
+* for codeblock syntax highlighting in markdown 
+*/
+
+pre .token.tag {
+	@apply text-green-300;
+}
+pre .token.attr-name {
+	@apply text-purple-300;
+}
+pre .token.attr-value {
+	@apply text-sky-300/90;
+}
+pre .token.namespace {
+	@apply text-purple-300;
+}
+pre .token.keyword,
+pre .token.operator {
+	@apply text-red-400;
+}
+pre .token.string {
+	@apply text-sky-300/90;
+}
+pre .token.function {
+	@apply text-sky-400;
+}
+pre .token.comment {
+	@apply text-gray-600;
+}
