+++ src/data/pages.ts
@@ -1,0 +1,27 @@
+import type { Page } from '$lib/types';
+
+const pages: Page[] = [
+	{
+		links: [
+			{
+				href: '',
+				text: 'Getting Started'
+			},
+			{
+				href: 'contributions',
+				text: 'Contributions'
+			}
+		]
+	},
+	{
+		title: 'Components',
+		links: [
+			{
+				href: 'sveltypicker',
+				text: 'SveltyPicker'
+			}
+		]
+	}
+];
+
+export { pages };
