+++ src/lib/i18n.ts
@@ -1,0 +1,16 @@
+import type {EN} from '$lib/types'
+
+export const en: EN = {
+    days:        ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'],
+    daysShort:   ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'],
+    daysMin:     ['Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su'],
+    months:      ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'],
+    monthsShort: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
+    meridiem:    ['am', 'pm'],
+    suffix:      ['st', 'nd', 'rd', 'th'],
+    todayBtn:    'Today',
+    clearBtn:    'Clear',
+    timeView:    'Show time view',
+    backToDate:  'Back to calendar view'
+}
+
