+++ src/lib/index.ts
@@ -1,0 +1,1 @@
+export { default as SveltyPicker } from './SveltyPicker/SveltyPicker.svelte';
