+++ src/routes/__layout.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+	import Navbar from '../components/Navbar.svelte';
+	import '../app.css';
+</script>
+
+<Navbar />
+<div id="page" class="mt-20 mb-40">
+	<slot />
+</div>
