+++ src/routes/docs/contributions.svx
@@ -1,0 +1,9 @@
+## 🏆 Thanks to:
+
+- [Bootstrap datepicker](https://github.com/smalot/bootstrap-datetimepicker/blob/master/js/bootstrap-datetimepicker.js) for some internal date and format handling
+
+PRs for extending built-in localization are welcome 🥳
+
+## Licence:
+
+MIT
