--- src/settings.js
@@ -1,17 +1,0 @@
-import { en } from './i18n.js';
-
-export default {
-  theme: 'sdt-calendar-colors',
-  mode: 'auto',
-  format: 'yyyy-mm-dd',
-  formatType: 'standard',
-  weekStart: 1,
-  visible: false,
-  inputClasses: null,
-  todayBtnClasses: 'sdt-action-btn sdt-today-btn',
-  clearBtnClasses: 'sdt-action-btn sdt-clear-btn',
-  todayBtn: true,
-  clearBtn: true,
-  autoclose: true,
-  i18n: en
-}
