+++ tailwind.config.cjs
@@ -1,0 +1,11 @@
+const config = {
+	content: ['./src/**/*.{html,js,svelte,ts}'],
+	darkMode: 'class',
+	theme: {
+		extend: {}
+	},
+
+	plugins: [require('@tailwindcss/typography')]
+};
+
+module.exports = config;
